Needing user input
Brand new to coding so go easy on me
I need to be able to input user information so that they can decide if they want to multiply divide add or subtract. I know I need to use the if statement and the else if but I am not sure how to make sure the user input is read (if this makes sense).